mqtt-weather-project/requirements.txt
paho-mqtt
rich
pytest
flake8
bandit
